requirements.txt
#--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.2.2
ctransformers[cuda]
datasets
accelerate
lxml
PyMuPDF
pyvi~=0.1.1
python-dotenv~=1.0.1
langchain~=0.2.5
langchain-community
langchain_huggingface
langchain-google-genai
langchain-chroma
langchain_rag
chromadb~=0.5.3
rank-bm25
streamlit~=1.35.0
PyMuPDF
pysqlite3-binary
beautifulsoup4~=4.12.3
sentence_transformers
requests~=2.32.3